+++ packages/dara-components/tests/python/common/test_button.py
@@ -26,7 +26,7 @@ class TestButtonComponent(unittest.TestCase):
                 'children': [Text(text='Click Here').dict(exclude_none=True)],
                 'stop_click_propagation': True,
                 'italic': False,
-                'onclick': action.dict(),
+                'onclick': action.dict(exclude_none=True),
                 'styling': ButtonStyle.PRIMARY.value,
                 'outline': False,
                 'position': 'relative',