+++ packages/dara-core/js/router/index.tsx
@@ -4,3 +4,4 @@ export { createRouter, findFirstPath } from './create-router';
 export { fetchRouteData, getFromPreloadCache, usePreloadRoute, type LoaderData } from './fetching';
 export { default as RouteContentDisplay, createRouteLoader, type LoaderResult } from './route-content';
 export { default as RouterRoot } from './router-root';
+export * from './utils';